+++ app/partials/sidebar.php
@@ -11,7 +11,7 @@ if ('columns-1' === ($column = \Jentil()->utilities->page->layout->column()) &&
  * Primary Sidebar
  */
 if (\is_active_sidebar($id = \Jentil()->setups['Sidebars\Primary']->id)) { ?>
-    <aside id="primary-sidebar" class="site-sidebar widget-area" itemscope itemtype="http://schema.org/WPSideBar">
+    <aside id="primary-sidebar" class="site-sidebar widget-area">
         <?php \dynamic_sidebar($id); ?>
     </aside><!-- #primary -->
 <?php }
@@ -26,7 +26,7 @@ if ('columns-3' !== $column &&
  * Secondary sidebar
  */
 if (\is_active_sidebar($id = \Jentil()->setups['Sidebars\Secondary']->id)) { ?>
-    <aside id="secondary-sidebar" class="site-sidebar widget-area" itemscope itemtype="http://schema.org/WPSideBar">
+    <aside id="secondary-sidebar" class="site-sidebar widget-area">
         <?php \dynamic_sidebar($id); ?>
     </aside><!-- #secondary -->
 <?php }