+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Set up JS module bundling with [rollup js](https://rollupjs.org)
 - Add `Procfile`
 
+### Removed
+- Remove HTML microdata
+
 ## 0.10.2 - 2019-04-30
 
 ### Added